+++ laokou-service/laokou-admin/laokou-admin-client/src/main/java/org/laokou/admin/api/MessagesServiceI.java
@@ -36,6 +36,6 @@ public interface MessagesServiceI {
 
 	Result<Datas<MessageCO>> unreadList(MessageUnreadListQry qry);
 
-	Result<Long> unreadCount(MessageUnreadCountGetQry qry);
+	Result<Integer> unreadCount(MessageUnreadCountGetQry qry);
 
 }
